Fix conversion warnings and errors

+++ Source/GUI/Button.cpp
@@ -43,7 +43,7 @@ void Button::handleEvent(sf::Event e, const sf::RenderWindow& window)
         case sf::Event::MouseButtonPressed:
             switch(e.mouseButton.button) {
                 case sf::Mouse::Left:
-                    if (m_button.getGlobalBounds().contains(pos.x, pos.y)) {
+                    if (m_button.getGlobalBounds().contains((float)pos.x, (float)pos.y)) {
                         m_function();
                     }
 
@@ -77,8 +77,8 @@ void Button::updateText()
     m_text.setOrigin(m_text.getGlobalBounds().width  / 2,
                      m_text.getGlobalBounds().height / 2);
 
-    m_text.move(m_button.getGlobalBounds().width  / 2,
-                m_button.getGlobalBounds().height / 2.5);
+    m_text.move(m_button.getGlobalBounds().width  / 2.0f,
+                m_button.getGlobalBounds().height / 2.5f);
 }
